fix: Alarm rings even when the pop-up is turned off
