+++ taggedUserAlarm.js
@@ -11,12 +11,9 @@ function alarmModal(icon, feedComments) {
     position: fixed;
     top: 0;
     left: 0;
-    width: 100%;
-    height: 100%;
     display: flex;
     align-items: center;
     justify-content: flex-start;
-    z-index: 9999;
   `;
 
   const shadowRoot = shadowHost.attachShadow({ mode: "open" });
